#195 move AlterAction to PlaysqlMigrationAlterAction
