+++ src/main/java/org/dbflute/intro/app/web/playsql/migration/alter/AlterSQLResult.java
@@ -13,7 +13,7 @@
  * either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-package org.dbflute.intro.app.web.alter;
+package org.dbflute.intro.app.web.playsql.migration.alter;
 
 import java.util.List;
