+++ src/main/java/org/dbflute/intro/app/web/playsql/migration/alter/PlaysqlMigrationAlterAction.java
@@ -18,16 +18,21 @@ package org.dbflute.intro.app.web.playsql.migration.alter;
 import javax.annotation.Resource;
 
 import org.dbflute.intro.app.logic.exception.DirNotFoundException;
-import org.dbflute.intro.app.logic.playsql.migration.PlaysqlMigrationLogic;
+import org.dbflute.intro.app.logic.playsql.migration.PlaysqlMigrationPhysicalLogic;
+import org.dbflute.intro.app.logic.playsql.migration.PlaysqlMigrationUpdateLogic;
+import org.dbflute.intro.app.logic.playsql.migration.info.PlaysqlMigrationInfoLogic;
+import org.dbflute.intro.app.logic.playsql.migration.info.ref.PlaysqlMigrationDirReturn;
 import org.dbflute.intro.app.web.base.IntroBaseAction;
+import org.dbflute.intro.bizfw.annotation.NotAvailableDecommentServer;
 import org.dbflute.intro.bizfw.tellfailure.OpenDirNotFoundException;
+import org.dbflute.intro.mylasta.action.IntroMessages;
 import org.lastaflute.web.Execute;
 import org.lastaflute.web.response.JsonResponse;
-import org.lastaflute.web.servlet.request.ResponseManager;
 
 /**
  * @author cabos
  * @author prprmurakami
+ * @author jflute
  */
 public class PlaysqlMigrationAlterAction extends IntroBaseAction {
 
@@ -35,20 +40,59 @@ public class PlaysqlMigrationAlterAction extends IntroBaseAction {
     //                                                                           Attribute
     //                                                                           =========
     @Resource
-    private PlaysqlMigrationLogic playsqlMigrationLogic;
+    private PlaysqlMigrationInfoLogic playsqlMigrationInfoLogic;
     @Resource
-    private ResponseManager responseManager;
+    private PlaysqlMigrationUpdateLogic playsqlMigrationUpdateLogic;
+    @Resource
+    private PlaysqlMigrationPhysicalLogic playsqlMigrationPhysicalLogic;
+    @Resource
+    private PlaysqlMigrationAlterAssist playsqlMigrationAlterAssist;
 
     // ===================================================================================
     //                                                                             Execute
     //                                                                             =======
+    /**
+     * Respond migration directory information of dbflute_client/playsql/migration.
+     * @param projectName name of client project (NotNull)
+     * @return migration directory information (NotNull)
+     */
     @Execute
-    public JsonResponse<Void> open(String clientName) {
+    public JsonResponse<AlterSQLResult> index(String projectName) {
+        PlaysqlMigrationDirReturn migrationDirReturn = playsqlMigrationInfoLogic.loadPlaysqlMigrationDir(projectName);
+        AlterSQLResult alterSQLResult = playsqlMigrationAlterAssist.mappingAlterSQLResult(migrationDirReturn);
+        return asJson(alterSQLResult);
+    }
+
+    @Execute
+    public JsonResponse<Void> open(String projectName) {
         try {
-            playsqlMigrationLogic.openAlterDir(clientName);
+            playsqlMigrationPhysicalLogic.openAlterDir(projectName);
         } catch (DirNotFoundException e) {
             throw new OpenDirNotFoundException("alter directory is not found. dirPath: " + e.getDirPath(), e.getDirPath());
         }
         return JsonResponse.asEmptyBody();
+    }
+
+    @Execute
+    @NotAvailableDecommentServer
+    public JsonResponse<Void> prepare(String projectName) {
+        playsqlMigrationUpdateLogic.unzipCheckedAlterZip(projectName);
+        playsqlMigrationUpdateLogic.copyUnreleasedAlterDir(projectName);
+        return JsonResponse.asEmptyBody();
+    }
+
+    @Execute
+    @NotAvailableDecommentServer
+    public JsonResponse<Void> create(String projectName, AlterCreateBody body) {
+        validate(body, messages -> moreValidate(projectName, body, messages));
+        playsqlMigrationUpdateLogic.createAlterSql(projectName, body.alterFileName);
+        return JsonResponse.asEmptyBody();
+    }
+
+    // ===================================================================================
+    //                                                                          Validation
+    //                                                                          ==========
+    private void moreValidate(String projectName, AlterCreateBody body, IntroMessages messages) {
+        playsqlMigrationAlterAssist.moreValidateCreate(projectName, body, messages);
     }
 }
